+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- Fix Two Factor Limit dialog layout
 - Don't use old string ids
 
 ## [0.1.3] - 2021-04-07