+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
+### Changed
+
+### Fixed
+
+## [0.1.5] - 2021-04-20
+### Added
 - Add disable all pins security setting
 - Add view with list of generated addresses
 - Support searching for a generated address